Revert "Rename theming to styling for better readability"
This reverts commit ea68499a42d4b2c0a6748199e5ae81adc6a4fdd9.
