+++ src/ert/gui/theming/manager.py
