+++ tests/ert/unit_tests/gui/theming/test_qss_loading.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import pytest
 
-from ert.gui.styling import ColorScheme
-from ert.gui.styling.theme_manager import theme as theme_module
-from ert.gui.styling.theme_manager.theme import load_qss
+from ert.gui.theming import ColorScheme
+from ert.gui.theming import theme as theme_module
+from ert.gui.theming.theme import load_qss
 
 
 class _MissingResource:
